+++ .eslintrc.js
@@ -1,5 +1,24 @@
 const restrictedGlobals = require('confusing-browser-globals');
 
+const restrictedSyntax = [
+  {
+    selector: 'ForStatement',
+    message: 'Consider using array iteration methods or for..of loops instead.',
+  },
+  {
+    selector: 'ForInStatement',
+    message: 'Consider using array iteration methods or for..of loops instead.',
+  },
+  {
+    selector: 'WhileStatement',
+    message: 'Consider using array iteration methods or for..of loops instead.',
+  },
+  {
+    selector: 'DoWhileStatement',
+    message: 'Consider using array iteration methods or for..of loops instead.',
+  },
+];
+
 /** @type {import('eslint').Linter.Config} */
 module.exports = {
   root: true,
@@ -15,7 +34,6 @@ module.exports = {
         'no-relative-import-paths',
         'import',
         'simple-import-sort',
-        'no-loops',
         'promise',
         'jasmine',
       ],
@@ -67,6 +85,7 @@ module.exports = {
         'simple-import-sort/imports': 'error',
         'simple-import-sort/exports': 'error',
         'no-restricted-globals': ['error'].concat(restrictedGlobals),
+        'no-restricted-syntax': ['error'].concat(restrictedSyntax),
         'promise/no-multiple-resolved': ['error'],
         'regexp/letter-case': [
           'error',
@@ -91,7 +110,6 @@ module.exports = {
         'regexp/sort-alternatives': ['error'],
         'regexp/sort-character-class-elements': ['error'],
         'regexp/use-ignore-case': ['error'],
-        'no-loops/no-loops': ['error'],
       },
       settings: {
         'import/parsers': {
@@ -120,11 +138,76 @@ module.exports = {
     {
       files: ['**/*.js'],
       parserOptions: {
-        ecmaVersion: 2021,
+        ecmaVersion: 2024,
       },
       extends: ['plugin:prettier/recommended'],
       rules: {
         'sort-imports': ['error'],
+      },
+    },
+    {
+      files: ['**/*.mjs'],
+      parserOptions: {
+        sourceType: 'module',
+        ecmaVersion: 2024,
+      },
+      plugins: [
+        'no-relative-import-paths',
+        'import',
+        'simple-import-sort',
+        'promise',
+      ],
+      extends: [
+        'plugin:import/recommended',
+        'plugin:import/typescript',
+        'plugin:promise/recommended',
+        'plugin:no-unsanitized/recommended-legacy',
+        'plugin:regexp/recommended',
+        'plugin:prettier/recommended',
+      ],
+      rules: {
+        'no-extra-boolean-cast': ['error'],
+        'no-var': ['error'],
+        'prefer-const': ['error'],
+        'import/first': 'error',
+        'import/consistent-type-specifier-style': 'error',
+        'import/newline-after-import': ['error', { count: 1 }],
+        'import/no-absolute-path': 'error',
+        'import/no-relative-packages': 'error',
+        'import/no-useless-path-segments': [
+          'error',
+          {
+            noUselessIndex: true,
+          },
+        ],
+        'simple-import-sort/imports': 'error',
+        'simple-import-sort/exports': 'error',
+        'no-restricted-globals': ['error'].concat(restrictedGlobals),
+        'no-restricted-syntax': ['error'].concat(restrictedSyntax),
+        'promise/no-multiple-resolved': ['error'],
+        'regexp/letter-case': [
+          'error',
+          {
+            caseInsensitive: 'lowercase',
+            unicodeEscape: 'lowercase',
+            hexadecimalEscape: 'lowercase',
+            controlEscape: 'uppercase',
+          },
+        ],
+        'regexp/no-contradiction-with-assertion': ['error'],
+        'regexp/no-control-character': ['error'],
+        'regexp/no-extra-lookaround-assertions': ['error'],
+        'regexp/no-misleading-capturing-group': ['error'],
+        'regexp/no-misleading-unicode-character': ['error'],
+        'regexp/no-missing-g-flag': ['error'],
+        'regexp/prefer-escape-replacement-dollar-char': ['error'],
+        'regexp/prefer-named-backreference': ['error'],
+        'regexp/prefer-named-capture-group': ['error'],
+        'regexp/prefer-named-replacement': ['error'],
+        'regexp/prefer-quantifier': ['error'],
+        'regexp/sort-alternatives': ['error'],
+        'regexp/sort-character-class-elements': ['error'],
+        'regexp/use-ignore-case': ['error'],
       },
     },
   ],