+++ src/app/services/spreadsheet.service.ts
@@ -21,11 +21,13 @@ import { ConfigService } from 'src/app/services/config.service';
 import { LoggingService } from 'src/app/services/logging.service';
 import { createLocaleAlternatives } from 'src/app/shared/util.locales';
 import {
+  createKeyValueList,
   createSlug,
   getDateFromString,
   getFullUrl,
 } from 'src/app/shared/utils';
 import { environment } from 'src/environments/environment';
+import { withoutTrailingSlash } from 'ufo';
 
 enum SheetName {
   config = 'Referral Page',
@@ -44,8 +46,8 @@ export class SpreadsheetService {
   private loggingService = inject(LoggingService);
   private configService = inject(ConfigService);
 
-  static visibleKey = 'Show';
   static booleanTrueKey = 'Yes';
+  static hiddenValues = ['HIDE', 'HIDDEN', 'NO', 'N', 'FALSE', '0', '-'];
 
   static readCellValue(row: string[], key: number): string {
     if (!!row && !!row[key] && key < row.length) {
@@ -55,7 +57,11 @@ export class SpreadsheetService {
   }
 
   static isVisible(value: string): boolean {
-    return value === this.visibleKey;
+    if (this.hiddenValues.includes(value.toUpperCase())) {
+      return false;
+    }
+
+    return true;
   }
 
   static isBoolean(value: string): boolean {
@@ -64,20 +70,38 @@ export class SpreadsheetService {
 
   private getSheetUrl(region: string, sheetName: SheetName): string {
     const sheetId = this.configService.getRegionByRegionSlug(region)?.sheetId;
-    return `${environment.google_sheets_api_url}/${sheetId}/values/${sheetName}?key=${environment.google_sheets_api_key}&alt=json&prettyPrint=false`;
+    const apiUrl = withoutTrailingSlash(environment.google_sheets_api_url);
+    return `${apiUrl}/${sheetId}/values/${sheetName}?key=${environment.google_sheets_api_key}&alt=json&prettyPrint=false`;
   }
 
-  private getIndexOfTag(collection: string[], tagName: string) {
-    const tag = `#${tagName.toUpperCase()}`;
-    return collection.findIndex((value: string) => {
-      return value.toUpperCase().indexOf(tag) !== -1;
+  private getIndexOfTag({
+    collection,
+    tagName,
+  }: {
+    collection: string[];
+    tagName: string;
+  }) {
+    tagName = tagName.toUpperCase();
+    const index = collection.findIndex((value: string) => {
+      const collectionTag = value.split('#')[1];
+      if (!collectionTag) {
+        return false;
+      }
+      return collectionTag.toUpperCase() === tagName;
     });
+    return index;
   }
 
-  private createKeyMap(keys: string[], collection: string[]): KeyMap {
+  private createKeyMap({
+    tagNames,
+    collection,
+  }: {
+    tagNames: string[];
+    collection: string[];
+  }): KeyMap {
     const keyMap: KeyMap = new Map();
-    keys.forEach((keyName: string) => {
-      keyMap.set(keyName, this.getIndexOfTag(collection, keyName));
+    tagNames.forEach((tagName: string) => {
+      keyMap.set(tagName, this.getIndexOfTag({ collection, tagName }));
     });
     return keyMap;
   }
@@ -195,10 +219,10 @@ export class SpreadsheetService {
           );
         }
         const headerRow = response.values[0];
-        const categoriesColumnMap = this.createKeyMap(
-          Object.values(CategoryCol),
-          headerRow,
-        );
+        const categoriesColumnMap = this.createKeyMap({
+          tagNames: Object.values(CategoryCol),
+          collection: headerRow,
+        });
 
         return response.values
           .slice(1) // Remove header-row
@@ -264,10 +288,10 @@ export class SpreadsheetService {
           );
         }
         const headerRow = response.values[0];
-        const subCategoryColumnMap = this.createKeyMap(
-          Object.values(SubCategoryCol),
-          headerRow,
-        );
+        const subCategoryColumnMap = this.createKeyMap({
+          tagNames: Object.values(SubCategoryCol),
+          collection: headerRow,
+        });
 
         return response.values
           .slice(1) // Remove header-row
@@ -376,10 +400,10 @@ export class SpreadsheetService {
           );
         }
         const headerRow = response.values[0];
-        const offerColumnMap = this.createKeyMap(
-          Object.values(OfferCol),
-          headerRow,
-        );
+        const offerColumnMap = this.createKeyMap({
+          tagNames: Object.values(OfferCol),
+          collection: headerRow,
+        });
 
         return response.values
           .slice(1) // Remove header-row
@@ -670,20 +694,21 @@ export class SpreadsheetService {
         }
         const regionRows = response.values;
         const headerRow = regionRows[0];
-        const regionColumnMap = this.createKeyMap(
-          Object.values(RegionCol),
-          headerRow,
-        );
+        const regionColumnMap = this.createKeyMap({
+          tagNames: Object.values(RegionCol),
+          collection: headerRow,
+        });
         const valueCol = this.getIndexOrFallback(
           regionColumnMap.get(RegionCol.value),
           1,
         );
 
         const keysCol = this.createKeysCollection(regionRows, regionColumnMap);
-        const regionRowMap = this.createKeyMap(
-          Object.values(RegionDataKey),
-          keysCol,
-        );
+
+        const regionRowMap = this.createKeyMap({
+          tagNames: Object.values(RegionDataKey),
+          collection: keysCol,
+        });
 
         return this.convertConfigSheetToRegionData(
           regionRows,
@@ -755,7 +780,11 @@ export class SpreadsheetService {
           ? LoggingEvent.NotFoundParentQuestionIsSelf
           : LoggingEvent.NotFoundParentQuestion;
       this.loggingService.logEvent(LoggingEventCategory.error, errorType, {
-        name: `row=${element.id};slug=${element.slug};parent=${element.parentSlug}`,
+        name: createKeyValueList({
+          row: element.id,
+          slug: element.slug,
+          parent: element.parentSlug,
+        }),
         row: element.id,
         slug: element.slug,
         parentSlug: element.parentSlug,
@@ -783,7 +812,10 @@ export class SpreadsheetService {
           );
         }
         const headerRow = response.values[0];
-        const qaColumnMap = this.createKeyMap(Object.values(QACol), headerRow);
+        const qaColumnMap = this.createKeyMap({
+          tagNames: Object.values(QACol),
+          collection: headerRow,
+        });
 
         return response.values
           .slice(1) // Remove header-row